refactor: move tournament pages to /pages folder
