--- src/app/modules/tournament-management/components/tournament/tournament/tournament.component.html
@@ -1,65 +1,0 @@
-<form [formGroup]="validationForm">
-	<mat-tab-group>
-		<mat-tab label="General">
-			<!-- <ng-template matTabContent>
-				<app-tournament-general [tournament]="tournament" [validationForm]="validationForm"></app-tournament-general>
-			</ng-template> -->
-
-			Updated
-		</mat-tab>
-
-		<mat-tab label="wyBin">
-			<!-- <ng-template matTabContent>
-				<app-tournament-wybin [tournament]="tournament"></app-tournament-wybin>
-			</ng-template> -->
-
-			Updated
-		</mat-tab>
-
-		<mat-tab label="Access">
-			<!-- <ng-template matTabContent>
-				<app-tournament-access [tournament]="tournament"></app-tournament-access>
-			</ng-template> -->
-
-			Updated
-		</mat-tab>
-
-		<mat-tab label="Webhook">
-			<!-- <ng-template matTabContent>
-				<app-tournament-webhook [tournament]="tournament" [validationForm]="validationForm"></app-tournament-webhook>
-			</ng-template> -->
-
-			Updated
-		</mat-tab>
-
-		<mat-tab label="Stages">
-			<!-- <ng-template matTabContent>
-				<app-tournament-stages [tournament]="tournament" [validationForm]="validationForm"></app-tournament-stages>
-			</ng-template> -->
-
-			Updated
-		</mat-tab>
-
-		<mat-tab label="Conditional messages">
-			<!-- <ng-template matTabContent>
-				<app-tournament-conditional-message [tournament]="tournament" [validationForm]="validationForm"></app-tournament-conditional-message>
-			</ng-template> -->
-
-			Updated
-		</mat-tab>
-
-		<mat-tab label="Participants">
-			<!-- <ng-template matTabContent>
-				<app-tournament-participants [tournament]="tournament" [validationForm]="validationForm"></app-tournament-participants>
-			</ng-template> -->
-
-			Updated
-		</mat-tab>
-
-		<mat-tab label="Mappool">
-			<ng-template matTabContent>
-				<app-mappool-overview [tournament]="tournament" [validationForm]="validationForm"></app-mappool-overview>
-			</ng-template>
-		</mat-tab>
-	</mat-tab-group>
-</form>
