--- src/app/modules/tournament-management/components/tournament/tournament/tournament.component.scss
@@ -1,1 +1,0 @@
-
