--- src/app/modules/tournament-management/components/tournament/tournament/tournament.component.spec.ts
@@ -1,25 +1,0 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { TournamentComponent } from './tournament.component';
-
-describe('TournamentComponent', () => {
-	let component: TournamentComponent;
-	let fixture: ComponentFixture<TournamentComponent>;
-
-	beforeEach(async(() => {
-		TestBed.configureTestingModule({
-			declarations: [TournamentComponent]
-		})
-			.compileComponents();
-	}));
-
-	beforeEach(() => {
-		fixture = TestBed.createComponent(TournamentComponent);
-		component = fixture.componentInstance;
-		fixture.detectChanges();
-	});
-
-	it('should create', () => {
-		expect(component).toBeTruthy();
-	});
-});
