+++ src/app/modules/tournament-management/pages/tournament/tournament-access/tournament-access.component.scss
