+++ src/app/modules/tournament-management/pages/tournament/tournament-general/tournament-general.component.scss
