+++ src/app/modules/tournament-management/pages/tournament/tournament-mappool/mappool-overview/mappool-overview.component.html
