+++ src/app/modules/tournament-management/pages/tournament/tournament-mappool/mappool/mappool.component.scss
