+++ src/app/modules/tournament-management/pages/tournament/tournament-mappool/mod-bracket/mod-bracket.component.html
