+++ src/app/modules/tournament-management/pages/tournament/tournament-participants/tournament-participants.component.scss
