+++ src/app/modules/tournament-management/pages/tournament/tournament-stages/tournament-stages.component.ts
