+++ src/app/modules/tournament-management/pages/tournament/tournament-trigger-message/tournament-trigger-message.component.html
