+++ src/app/modules/tournament-management/pages/tournament/tournament-webhook/tournament-webhook.component.scss
