+++ src/app/modules/tournament-management/tournament-management-routing.module.ts
@@ -7,14 +7,14 @@ import { TournamentCreateComponent } from './pages/tournament-create/tournament-
 import { TournamentEditComponent } from './pages/tournament-edit/tournament-edit.component';
 import { LocalTournamentsComponent } from './pages/local-tournaments/local-tournaments.component';
 import { PublishedTournamentsComponent } from './pages/published-tournaments/published-tournaments.component';
-import { TournamentGeneralComponent } from './components/tournament/tournament-general/tournament-general.component';
-import { TournamentWybinComponent } from './components/tournament/tournament-wybin/tournament-wybin.component';
-import { TournamentAccessComponent } from './components/tournament/tournament-access/tournament-access.component';
-import { TournamentWebhookComponent } from './components/tournament/tournament-webhook/tournament-webhook.component';
-import { TournamentStagesComponent } from './components/tournament/tournament-stages/tournament-stages.component';
-import { TournamentParticipantsComponent } from './components/tournament/tournament-participants/tournament-participants.component';
-import { MappoolOverviewComponent } from './components/tournament/tournament-mappool/mappool-overview/mappool-overview.component';
-import { TournamentTriggerMessageComponent } from './components/tournament/tournament-trigger-message/tournament-trigger-message.component';
+import { TournamentGeneralComponent } from './pages/tournament/tournament-general/tournament-general.component';
+import { TournamentWybinComponent } from './pages/tournament/tournament-wybin/tournament-wybin.component';
+import { TournamentAccessComponent } from './pages/tournament/tournament-access/tournament-access.component';
+import { TournamentWebhookComponent } from './pages/tournament/tournament-webhook/tournament-webhook.component';
+import { TournamentStagesComponent } from './pages/tournament/tournament-stages/tournament-stages.component';
+import { TournamentParticipantsComponent } from './pages/tournament/tournament-participants/tournament-participants.component';
+import { MappoolOverviewComponent } from './pages/tournament/tournament-mappool/mappool-overview/mappool-overview.component';
+import { TournamentTriggerMessageComponent } from './pages/tournament/tournament-trigger-message/tournament-trigger-message.component';
 
 const routes: Routes = [
 	{
